+++ src/main/java/com/example/jobis/domain/company/presentation/dto/response/CompanyMyPageResponse.java
@@ -1,4 +1,4 @@
-package com.example.jobis.domain.company.controller.dto.response;
+package com.example.jobis.domain.company.presentation.dto.response;
 
 import com.example.jobis.domain.company.domain.enums.CompanyType;
 import lombok.Builder;
