+++ src/main/java/com/example/jobis/domain/company/service/RegisterCompanyService.java
@@ -2,7 +2,7 @@ package com.example.jobis.domain.company.service;
 
 import com.example.jobis.domain.company.domain.repository.CompanyRepository;
 import com.example.jobis.domain.user.controller.dto.response.TokenResponse;
-import com.example.jobis.domain.company.controller.dto.request.RegisterCompanyRequest;
+import com.example.jobis.domain.company.presentation.dto.request.RegisterCompanyRequest;
 import com.example.jobis.domain.company.domain.Company;
 import com.example.jobis.domain.company.exception.CompanyAlreadyExistsException;
 import com.example.jobis.domain.company.exception.CompanyNotFoundException;
